EXOSCALE: update client lib to v0.59.0 (#1183)
REGRESSION and BUILD fix.  dnscontrol compiles with this but I can't test the
EXOSCALE behavior to be sure it's logically sound.

The dependency was v1.19.0 but this git tag has been withdrawn and the go.mod
for that git repository contains:

    retract v1.19.0 // Published accidentally.

Per <https://proxy.golang.org/github.com/exoscale/egoscale/@v/v1.19.0.info> the
timestamp for the bogus v1.19.0 dependency is 2019-10-29, which corresponds to
v0.19.0.

On 2020-02-27, the dnscontrol repo updated this dependency to v0.23.0, so it
now being v1.19.0 (== v0.19.0) is a REGRESSION.

I suggest that the maintainers purge this bogus v1.19.0 from their systems to
try to prevent it leaking in again in future.