Add calling an executable for credentials.
Certain password managers include a CLI that allows for injecting
credentials into a template file. Using `!` works for a single shell
script, but doesn't work so well for calling a program, considering it
explicitly ignores items in the path. This allows for tools like
1Password CLI to be used to inject the secrets into the json,
without ever needing to write the credentials to disk.