AXFRDDNS: fix #1913
The integration tests do not fail when executed with a `bind` server,
but they fail when executedwith a `knot` server.

It looks like knot is buggy and does not allow to replace a `CNAME`
record by a `A` record in a single batched update.

This patch adds an option `buggy-cname` in `creds.json` for splitting
such request in two DDNS updates. This option is deactivated by
default, since the update is not atomic anymore.