AXFRDDNS: allow to modify the last NS record of a zone
Servers should silently ignore the removal of the last NS record of a
zone. Then, when updating a NS record, we cannot safely remove the old
record before inserting the new one, as we do for all other record's
kinds, or when there is more that one NS record.

For preserving this simple scheme, when a update requires to modify a
NS record, we first insert a dummy NS record. This record is removed
at the end of the same set of batched updates. Therefore, it will
never be commited by the server in the zone and never served by the
server to a client.