NS1: Handle assigned nameservers better
It was discovered in #2217 that, as of recently, NS1 may return nameservers
with a trailing dot. Attempting to fix, we discovered that this creates
another bug, #2254. This is triggered when NS1 returns entries without a dot,
which happens when new zones are created.

Dnscontrol is opinionated and always adds trailing dots to nameservers.

So, instead of picking one out of models.ToNameservers/ToNameserversStripTD()
and breaking one case every time, let's add some code that covers both cases:
* If dots exist, remove them and then submit the clean array to ToNameservers().
* If not, just pass the original nameserver to ToNameservers().

Dnscontrol will then ensure that trailing-dotted versions are configured into
NS1 for all managed domains and remove this inconsistency.

Fixes #2254