Disable ansi colors in notifications
Not very long ago, ansi color support was enabled in dnscontrol, to
ease reviewing of changesets.

This doesn't play well with notifications, as they usually can't
render these codes; slack for example outputs the escape codes,
leading to ugly and hard to grasp messages.

This changeset ensures that any message coming to a notifier will
get stripped of ansi color codes, before passed on to the notifier.

That way, notifiers are then free to re-add colors in a way that is
supported in the relevant platform.