Fix integration testcase 19
Integration testcase 19 contains a MX record which refers to an A
record with the same domain as the MX record itself. So it seems
like a self-reference, while it shouldn't be treated as such.

The algorithm has been changed to not find self-references at all,
in DNS world these do not exist anyway. The algorithm also ignores
duplicate references to the same record, this is more of a side
effect, but the algorithm only needs one reference anyway. And at
last it will add edges to ALL referenced records if possible. The
previous approuch will only refer to one record instead of all.