Allow native OVH records to be managed
DKIM, SPF, DMARC OVH native records (such as when created in the UI) couldn't
be managed by dnscontrol natively because of various issue (for instance it is not possible
to update such record and change their type).
This implements native handling of those records when they already exist in the zone and manage them
as TXT record without changing their native types.
This patch also adds a bunch of integration tests specific to OVH to make sure we don't get any
regressions in the future.
To be noted that it is not possible to create such native OVH records (outside of what has been done
in the integration tests).