Allow ovh provider to connect to other endpoints than EU (#2625)
This introduces the `endpoint` provider configuration parameter.
If not provided the provider will connect to the EU region for
backward compatibility.
The `endpoint` can be set to `eu`, `us` or `ca` depending on the OVH
region the use wants to manage domains and zones in.
It can also be set to any arbitrary URL for future OVH regions.