FEATURE: CAA_BUILDER: add `issue_critical` and `issuewild_critical`
It is my understanding that - according to https://datatracker.ietf.org/doc/html/rfc8659 -
the critical flag can be applied to whatever tag of a CAA record.

This change of CAA_BUILDER adds the ability to provide this flag for the `issue` and
`issuewild` tags by setting respectively `issue_critical` and `issuewild_critical` to
`true`.