Ignore RRSets of type REDIRECT coming from NS1
NS1 replaces URLFWD with REDIRECTs. The latter supports HTTPS, which is
nice, but it is not a 'real' DNS Record from the configuration point of
view. To create a redirect, there is a separate API Endpoint. Creating a
redirect may also involve Certificate management, if a user wants HTTPS
support.
However, NS1 returns names, for which there is a redirect configured in
n.Zones.Get(), with a stub answers section 'NS1 MANAGED RECORD'.
RRType of REDIRECT confuses dnscontrol, and entire sync fails.

This commit ignores REDIRECTs coming from NS1 API.