AXFRDDNS: Remove `buggy-cname`
It's semantically invalid for a CNAME record to coexist with other
(non-DNSSEC) records, and RFC 2136 §3.4.2.2 requires DDNS recipients to
ignore any such updates. Therefore, it should always be safe to
unconditionally delete any records for a given name before we add a
CNAME record for that name.

`buggy-cname` was only needed as a workaround for a Knot DNS bug/feature
that caused it to ignore an additional followed by a deletion to the
same CNAME record. However, if we delete *all* records for a given name
(instead of just the CNAME records), Knot DNS will process the update as
expected. Therefore, we no longer need `buggy-cname`.

This has only been tested with Knot DNS, but given the semantics of
CNAME records, this _should_ work with any RFC-compliant DNS server.

Fixes #3292.