AXFRDDNS: Change the default value of `update-mode` to `tcp`
AXFR is only defined over TCP, and DNSControl always requests an AXFR
before sending an UPDATE, therefore any server supported by DNSControl
is already capable of handling TCP connections. The only advantage to
using UDP for DNS is that it has lower latency, but that's only relevant
for queries, not updates. By using TCP for updates, we should be more
resilient to network issues, and we should be able to handle larger
updates.

(Maybe?) fixes #2581.