Changes to '--domains' flag processing
This changes the pattern matching of `--domains=` to use the unique name of
each domain, ensuring that tagged domains are matched when specified in the
filter. Additionally, wildcard matching of tags is now supported, making
`--domains=example.com!*` match any tagged `example.com` DomainConfig.

New tests for whichZonesToProcess and additional cases for domainInList are
included.

Updated documentation for D and preview/push.