AUTODNS: enable concurrent gathering of zones
Besides enabling the feature toggle this patch adds proper error propagation to all internal method calls to mitigate crashing if an API request fails.

When working with a large amount of zones one might run into a rate limit, if this happens the provider will retry the request after a slight delay.