INWX: Fix Null MX handling
This adds special handling of MX records for INWX. The API rejects
record updates that violate RFC 7505: if a Null MX RR is present in
the zone, new regular MX records are rejected. The inverse is also
rejected - if MX records exist in the zone, a Null MX record can not
be added.

Separately from the Null MX issue, changes to MX preference from
non-zero to zero are accepted but not applied, leading to inconsistent
state between the dnscontrol configuration and the published zone.