BUILD: Add version package
Commit 9873f9f4b2fece7a46053eaa375fe24070971e09 (v4.7.4) removed the
version package in favour of using GoReleaser to set the version, which
it only sets in main and is not passed down to other packages.

This means that if any other package, notably two providers which use
the version for setting a User-Agent, wants access to the version, they
no longer have it.

This change reintroduces the version package, setting GoReleaser to set
the version here, and refactors all uses of the version to use the
package.

Note that if you run `go run main.go version` the version output by Go
for the main module is "(devel)". Using build/build.go gives the commit
hash.