DNSIMPLE: mark CanConcur
The `dnsimpleProvider` is a simple object of 3 strings with no state.

Using `dnscontrol preview --cmode all` in a configuration with 12 DNSimple
zones, and the binary built with race detection, results in no warnings.

Fixed `concurrency-verified.md` to state "the provider in question" instead of
an early provider to support this.  (Also some end-of-line cleanups for
Markdown; trailing whitespace is semantically significant in block mode but not
AFAIK for lists, so I think this is a safe change; this was my text-editor
honoring the project's `.editorconfig` settings, which don't have an EOL
whitespace exemption for markdown).

Index doc regenerated.