DNSIMPLE: pedantic correctness fix for concurrency
While updating the docs on what CanConcur means, I realised that the
`.accountID` field in `dnsimpleProvider` is a cache and subject to races.  It
will always be assigned the same value, and a `-race` build didn't catch it,
but it wasn't strictly correct to just declare this provider to be
concurrency-safe.

So use `sync.Once` to properly protect the account ID field update.