NEW RECORD TYPE: OPENPGPKEY (basic)
This is a follow-up to the efforts in:
https://github.com/StackExchange/dnscontrol/issues/2370
https://github.com/StackExchange/dnscontrol/pull/2374

Whilst full parsing and conversion of address and key material would be
preferable and more in line with the ideals of dnscontrol, the logic
needed requires significant development effort.

In order not to further completely prevent zones containing OPENPGPKEY
records from being managed by dnscontrol, add basic support for the
record type which enables a 1:1 mapping of OPENPGPKEY records from
dnscontrol to enabled providers.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>