PORKBUN: Configurable retry parameters
The Porkbun provider can now configure the API retry mechanism by
setting max_attempts and/or max_duration in the credentials
configuration. This aims to help users who have many domains configured
and run in to Porkbun API limits with the current retry mechanism.

The max_attempts setting controls how many attempts (in total, not
retries) will be made for each domain. The default is 5.

The max_duration allows the retry mechanism to keep retrying up to this
duration, for each domain.

The current behaviour is to do exponential backoff between 1 second and
10 seconds, with 5 retries (6 total attempts), with jitter of +/-100ms
(max, it will vary). The jitter means the provider will often retry
sooner than 1 second, which the Porkbun API doesn't like. To counter
this, the minimum retry time is now 1.2 seconds.

With the new behaviour the max backoff time is still 10 seconds, but if
a long max_duration is set it allows the provider to keep retrying every
10 seconds.