CHORE: Update the link to future video meetings (#3917)
# Issue

GItHub Discussions can only have 1 category. With a separate category
for the meeting announcements, it can't also appear in the
announcements, which is where people look for it. This lead to bad
user-experience.

# Resolution

Use a label instead, and post the announcements to the "Announcements"
category. Update the link to use labels.