NEW FEATURE: Make "fmt" more like other formatters (warning: New default for -o) (#3927)
# Issue

The "fmt" command doesn't work like Terraform fmt, gofmt, and others. It
should:

* Default should DTRT: Format the existing dnsconfig.js file in place
* Not output a filename if no changes were made
* Don't modify the file's timestamp if no changes were made
* Accept stdin and stdout

# Resolution

All that.

# Potential breaking change

The default for '-o' is now dnsconfig.js instead of stdout. This may
break shell scripts that redirect stdout instead of using '-o' to
specify an output filename. If you use ">" instead of "-o", please
switch to "-o".