FIX: Make type definition for NewDnsProvider support meta properly
Prior to this it would only work if you did the deprecated style of
passing the provider type first.

```js
NewDnsProvider("bind", "BIND" {"stuff": "stuff"})
```

But it wouldn't support omitting the optional provider type and
specifying the extra object.

```js
NewDnsProvider("bind", {"stuff": "stuff"})
```

Now the old style is forbidden (in the type hints) to discourage use of
it, although dnscontrol does still allow it.

Fixes #3946